fix calendar event webhook
